token

+++ app/Http/Controllers/ExpLvlController.php
@@ -15,26 +15,50 @@ use Illuminate\Http\Request;
 class ExpLvlController extends Controller
 {
     public function createLvl(Request $request){
-        $lvl = ExpLvl::create($request->all());
-        return response()->json($lvl);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $lvl = ExpLvl::create($request->all());
+            return response()->json($lvl);
+        }else{
+            return response()->json('The token does not match');
+        }
+
     }
 
     public function updateLvl(Request $request, $id){
-        $lvl  = ExpLvl::find($id);
-        $lvl->make = $request->input('make');
-        $lvl->save();
-        return response()->json($lvl);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $lvl  = ExpLvl::find($id);
+            $lvl->make = $request->input('make');
+            $lvl->save();
+            return response()->json($lvl);
+        }else{
+            return response()->json('The token does not match');
+        }
+
     }
 
-    public function deleteLvl($id){
-        $lvl  = ExpLvl::find($id);
-        $lvl->delete();
-        return response()->json('Removed successfully.');
+    public function deleteLvl(Request $request, $id){
+        $token = $request->input('token');
+        if($token == $this->token){
+            $lvl  = ExpLvl::find($id);
+            $lvl->delete();
+            return response()->json('Removed successfully.');
+        }else{
+            return response()->json('The token does not match');
+        }
+
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $lvl = ExpLvl::all();
-        return response()->json($lvl);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $lvl = ExpLvl::all();
+            return response()->json($lvl);
+        }else{
+            return response()->json('The token does not match');
+        }
+
     }
 }
