+++ app/Http/Controllers/ProjectController.php
@@ -15,27 +15,55 @@ use Illuminate\Http\Request;
 class ProjectController extends Controller
 {
     public function createProject(Request $request){
-        $input = $request->all();
-        $project = Project::create($input);
-        return response()->json($project);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $input = $request->all();
+            $project = Project::create($input);
+            return response()->json($project);
+        }else{
+            return response()->json('The token does not match');
+        }
+
+
     }
 
     public function updateProject(Request $request, $id){
-        $project  = Project::find($id);
-        $project->make = $request->input('make');
-        $project->save();
-        return response()->json($project);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $project  = Project::find($id);
+            $project->make = $request->input('make');
+            $project->save();
+            return response()->json($project);
+        }else{
+            return response()->json('The token does not match');
+        }
+
+
     }
 
-    public function deleteProject($id){
-        $project  = Project::find($id);
-        $project->delete();
-        return response()->json('Removed successfully.');
+    public function deleteProject(Request $request, $id){
+        $token = $request->input('token');
+        if($token == $this->token){
+            $project  = Project::find($id);
+            $project->delete();
+            return response()->json('Removed successfully.');
+        }else{
+            return response()->json('The token does not match');
+        }
+
+
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $project = Project::all();
-        return response()->json($project);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $project = Project::all();
+            return response()->json($project);
+        }else{
+            return response()->json('The token does not match');
+        }
+
+
     }
 }
