+++ app/Http/Controllers/UserGroupController.php
@@ -14,27 +14,54 @@ use Illuminate\Http\Request;
 
 class UserGroupController extends Controller
 {
-    public function createUserGroup(Request $request){
-        $userGroup = UserGroup::create($request->all());
-        return response()->json($userGroup);
-    }
-
-    public function updateUserGroup(Request $request, $id){
-        $userGroup  = UserGroup::find($id);
-        $userGroup->make = $request->input('make');
-        $userGroup->save();
-        return response()->json($userGroup);
-    }
-
-    public function deleteUserGroup($id){
-        $userGroup  = UserGroup::find($id);
-        $userGroup->delete();
-        return response()->json('Removed successfully.');
-    }
-
-    public function index()
+    public function createUserGroup(Request $request)
     {
-        $userGroup = UserGroup::all();
-        return response()->json($userGroup);
+        $token = $request->input('token');
+        if($token == $this->token){
+            $userGroup = UserGroup::create($request->all());
+            return response()->json($userGroup);
+        }else{
+            return response()->json('The token does not match');
+        }
+
+    }
+
+    public function updateUserGroup(Request $request, $id)
+    {
+        $token = $request->input('token');
+        if($token == $this->token){
+            $userGroup  = UserGroup::find($id);
+            $userGroup->make = $request->input('make');
+            $userGroup->save();
+            return response()->json($userGroup);
+        }else{
+            return response()->json('The token does not match');
+        }
+
+    }
+
+    public function deleteUserGroup(Request $request, $id)
+    {
+        $token = $request->input('token');
+        if($token == $this->token){
+            $userGroup  = UserGroup::find($id);
+            $userGroup->delete();
+            return response()->json('Removed successfully.');
+        }else{
+            return response()->json('The token does not match');
+        }
+
+    }
+
+    public function index(Request $request)
+    {
+        $token = $request->input('token');
+        if($token == $this->token){
+            $userGroup = UserGroup::all();
+            return response()->json($userGroup);
+        }else{
+            return response()->json('The token does not match');
+        }
+
     }
 }
